Make the status message more visible
The status message used to appear under the buttons. This was a bit
of a problem because one would have to scroll down in order to see
the message, and those messages could be lost.

So, instead, place the status message besides the buttons, to make
sure they're both more visible, and close to where the user's eyes
probably are anyway after clicking on the Save button.