function parseViewCount return fix
Fixing the return value of function parseViewCount. If a video had more than 10k view, the function would return only 10 instead of 10000 for example. It is because the views a stored with a space between then, and the function does not assume there will be a space between the numbers.